challenges/Web/babep1ckle/src/src/requirements.txt
Flask==2.2.5
PyJWT==2.8.0